container: Rework the children moving model
Until now, containers children were moving themselves and the
GESContainer class was listening to their notify signals to notice
it and start moving itself. This was introducing some state management
to be able to know why the child was moving and introduced quite a lot
of hard to manage side effects.

The new model is to have the children get call the parent moving
methods as needed letting the parent handle all the children move
without having to keep any internal state of operations.