composition: Output one single segment "per seek"
Instead of pushing a segment per new stack, output one single segment
that represents the whole duration of the composition. We now
retimestamp everything inside that segment using buffer PTS running
time.

The main goal of doing so is to avoid having the encoder drainning on
new segments which, with some encoders, leads to the introduction of
a few "silent" samples, meaning hicups/blips/blops.

https://bugzilla.gnome.org/show_bug.cgi?id=796831